feat: refactor location handling by extracting map picker into dedicated components and updating CreateServiceScreen architecture
